+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-common/src/main/java/org/apache/shardingsphere/encrypt/algorithm/RC4EncryptAlgorithm.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.encrypt.strategy.impl;
+package org.apache.shardingsphere.encrypt.algorithm;
 
 import lombok.Getter;
 import lombok.Setter;
